Heal with cast button
This works basically fine while healing at max soul, if you lower the amount of soul needed to heal it will immediately cast a spell afterwards unless you hold down the focus button for a moment.

I guess that could be an upside or a downside sometimes.

This version of the mod would probably need an explainer or something added to let people know to set up quick cast.